seperated notifications in its own service
