+++ services/Notifications/notifications.js
@@ -1,0 +1,13 @@
+const all_notifications = () => {
+    const notfications = [
+        "Surah Mulk after Maghrib", 
+        "Dars after Zuhr", 
+        "Collections for Eid after Jummah", 
+        "Eid on the 23rd of July"
+    ]
+    return notfications
+}
+
+module.exports = {
+    "all_notifications" : all_notifications,
+}
